example/centralized_fl/server/requirements.txt
torch
torchvision
# scikit-learn
tqdm
# matplotlib
python-dotenv
schedule
# torchinfo
# requests
GitPython
numpy